requirements.txt
flask
google-generativeai
pandas
openpyxl
gunicorn